Move tests to `eth_test`
