+++ eth/eth_test.go
@@ -8,6 +8,7 @@ import (
 	opeth "github.com/ethereum-optimism/optimism/op-service/eth"
 	"github.com/ethereum/go-ethereum"
 	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/common/hexutil"
 	ethtypes "github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/rpc"
 	"github.com/polymerdao/monomer"
@@ -133,14 +134,87 @@ func TestGetBlockByHash(t *testing.T) {
 	}
 }
 
+// generateProofAPI creates a ProofAPI instance for testing
+func generateProofAPI(t *testing.T, blockStoreIsEmpty bool) *eth.ProofAPI {
+	t.Helper()
+	blockStore := testutils.NewLocalMemDB(t)
+	db := testutils.NewEthStateDB(t)
+
+	if !blockStoreIsEmpty {
+		block := testutils.GenerateBlock(t)
+		require.NoError(t, blockStore.AppendBlock(block))
+		require.NoError(t, blockStore.UpdateLabels(block.Header.Hash, block.Header.Hash, block.Header.Hash))
+	}
+
+	return eth.NewProofAPI(db, blockStore)
+}
+
 func TestGetProof(t *testing.T) {
-	someAddress := common.HexToAddress("0xabc")
-	blockstore := testutils.NewLocalMemDB(t)
-
-	proofAPI := eth.NewProofAPI(nil, blockstore)
-
 	blockNumber := rpc.LatestBlockNumber
-	pf, err := proofAPI.GetProof(context.Background(), someAddress, []string{}, rpc.BlockNumberOrHash{BlockNumber: &blockNumber})
-	require.Error(t, err, "should not succeed in generating proofs with empty blockstore")
-	require.Nil(t, pf, "received proof from empty blockstore")
+	zeroHash := common.Hash{}
+	zeroHashStr := zeroHash.String()
+	someAddress := common.HexToAddress("0xabc")
+	zeroBig := (*hexutil.Big)(new(big.Int).SetBytes(zeroHash.Bytes()))
+
+	testCases := []struct {
+		name              string
+		blockStoreIsEmpty bool
+		storageKeys       []string
+		expectedResult    *ethapi.AccountResult
+		expectError       bool
+	}{
+		{
+			name:              "empty blockstore",
+			blockStoreIsEmpty: true,
+			expectError:       true,
+		},
+		{
+			name:              "blockstore with block without storageKeys",
+			blockStoreIsEmpty: false,
+			expectedResult: &ethapi.AccountResult{
+				Address:      someAddress,
+				AccountProof: nil,
+				Balance:      zeroBig,
+				CodeHash:     zeroHash,
+				Nonce:        hexutil.Uint64(0),
+				StorageHash:  zeroHash,
+				StorageProof: []ethapi.StorageResult{},
+			},
+		},
+		{
+			name:              "blockstore with block with storageKeys and nil storageTrie",
+			blockStoreIsEmpty: false,
+			storageKeys:       []string{zeroHashStr},
+			expectedResult: &ethapi.AccountResult{
+				Address:      someAddress,
+				AccountProof: nil,
+				Balance:      zeroBig,
+				CodeHash:     zeroHash,
+				Nonce:        hexutil.Uint64(0),
+				StorageHash:  zeroHash,
+				StorageProof: []ethapi.StorageResult{
+					{
+						Key:   zeroHashStr,
+						Value: &hexutil.Big{},
+						Proof: []string{},
+					},
+				},
+			},
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			proofAPI := generateProofAPI(t, tc.blockStoreIsEmpty)
+			pf, err := proofAPI.GetProof(context.Background(), someAddress, tc.storageKeys, rpc.BlockNumberOrHash{BlockNumber: &blockNumber})
+
+			if tc.expectError {
+				require.Error(t, err, "should not succeed in generating proofs")
+				require.Nil(t, pf, "received proof when error was expected")
+			} else {
+				require.NoError(t, err, "should succeed in generating proofs")
+				require.Equal(t, tc.expectedResult, pf)
+			}
+		})
+	}
 }
